feat(frontend): auth store generic sign-in function
