+++ src/frontend/src/lib/stores/auth.store.ts
@@ -1,18 +1,10 @@
-import {
-	AUTH_MAX_TIME_TO_LIVE,
-	AUTH_POPUP_HEIGHT,
-	AUTH_POPUP_WIDTH,
-	INTERNET_IDENTITY_CANISTER_ID,
-	LOCAL_REPLICA_HOST
-} from '$lib/constants/app.constants';
-import { isDev } from '$lib/env/app.env';
 import { AuthBroadcastChannel } from '$lib/providers/auth-broadcast.provider';
 import { createAuthClient, safeCreateAuthClient } from '$lib/providers/auth-client.provider';
-import { SignInError, SignInInitError, SignInUserInterruptError } from '$lib/types/errors';
+import type { SignInFn } from '$lib/types/auth';
+import { SignInInitError } from '$lib/types/errors';
 import type { OptionIdentity } from '$lib/types/itentity';
 import type { Option } from '$lib/types/utils';
-import { popupCenter } from '$lib/utils/window.utils';
-import { type AuthClient, ERROR_USER_INTERRUPT } from '@dfinity/auth-client';
+import { type AuthClient } from '@dfinity/auth-client';
 import { isNullish, nonNullish } from '@dfinity/utils';
 import { type Readable, writable } from 'svelte/store';
 
@@ -25,7 +17,7 @@ let authClient: Option<AuthClient>;
 export interface AuthStore extends Readable<AuthStoreData> {
 	sync: () => Promise<void>;
 	forceSync: () => Promise<void>;
-	signInWithII: () => Promise<void>;
+	signIn: (params: { signInFn: SignInFn }) => Promise<void>;
 	signOut: () => Promise<void>;
 }
 
@@ -73,54 +65,33 @@ const initAuthStore = (): AuthStore => {
 			await sync({ forceSync: true });
 		},
 
-		signInWithII: () =>
-			// eslint-disable-next-line no-async-promise-executor
-			new Promise<void>(async (resolve, reject) => {
-				if (isNullish(authClient)) {
-					reject(new SignInInitError());
-					return;
+		signIn: async ({ signInFn }) => {
+			if (isNullish(authClient)) {
+				throw new SignInInitError();
+			}
+
+			const { identity } = await signInFn({ authClient });
+
+			set({ identity });
+
+			const broadCastSignIn = () => {
+				try {
+					// If the user has more than one tab open in the same browser,
+					// there could be a mismatch of the cached delegation chain vs the identity key of the `authClient` object.
+					// This causes the `authClient` to be unable to correctly sign calls, raising Trust Errors.
+					// To mitigate this, we use a BroadcastChannel to notify other tabs when a login has occurred, so that they can sync their `authClient` object.
+					const bc = new AuthBroadcastChannel();
+					bc.postLoginSuccess();
+				} catch (err: unknown) {
+					// We don't really care if the broadcast channel fails to open or if it fails to post messages.
+					// This is a non-critical feature that improves the UX when OISY is open in multiple tabs.
+					// We just print a warning in the console for debugging purposes.
+					console.warn('Auth BroadcastChannel posting failed', err);
 				}
+			};
 
-				const identityProvider = isDev()
-					? /apple/i.test(navigator?.vendor)
-						? `${LOCAL_REPLICA_HOST}?canisterId=${INTERNET_IDENTITY_CANISTER_ID}`
-						: `http://${INTERNET_IDENTITY_CANISTER_ID}.${new URL(LOCAL_REPLICA_HOST).host}`
-					: `https://identity.internetcomputer.org`;
-
-				await authClient?.login({
-					maxTimeToLive: AUTH_MAX_TIME_TO_LIVE,
-					allowPinAuthentication: false,
-					onSuccess: () => {
-						set({ identity: authClient?.getIdentity() });
-
-						try {
-							// If the user has more than one tab open in the same browser,
-							// there could be a mismatch of the cached delegation chain vs the identity key of the `authClient` object.
-							// This causes the `authClient` to be unable to correctly sign calls, raising Trust Errors.
-							// To mitigate this, we use a BroadcastChannel to notify other tabs when a login has occurred, so that they can sync their `authClient` object.
-							const bc = new AuthBroadcastChannel();
-							bc.postLoginSuccess();
-						} catch (err: unknown) {
-							// We don't really care if the broadcast channel fails to open or if it fails to post messages.
-							// This is a non-critical feature that improves the UX when OISY is open in multiple tabs.
-							// We just print a warning in the console for debugging purposes.
-							console.warn('Auth BroadcastChannel posting failed', err);
-						}
-
-						resolve();
-					},
-					onError: (error?: string) => {
-						if (error === ERROR_USER_INTERRUPT) {
-							reject(new SignInUserInterruptError(error));
-							return;
-						}
-
-						reject(new SignInError(error));
-					},
-					identityProvider,
-					windowOpenerFeatures: popupCenter({ width: AUTH_POPUP_WIDTH, height: AUTH_POPUP_HEIGHT })
-				});
-			}),
+			broadCastSignIn();
+		},
 
 		signOut: async () => {
 			const client: AuthClient = authClient ?? (await createAuthClient());
